requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-cors-headers==2.3.0
django-filter==1.1.0
django-location-field==2.0.3
django-multiselectfield==0.1.8
django-webpack-loader==0.5.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
Pillow==5.0.0
PyJWT==1.6.4
pytz==2017.3
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
yolk3k==0.9
